feat: CarWinner 클래스 추가
